# src/UI/FluentCMS.Web.ApiClients/swagger.json
{
  "openapi": "3.0.4",
  "info": {
    "title": "FluentCMS API",
    "version": "v1.0.0"
  },
  "paths": {
...
    "/api/Folders/GetRoot": {
      "get": {
        "tags": [
          "Folders"
        ],
        "operationId": "GetRoot",
        "parameters": [
          {
            "name": "siteId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Folders/GetParentFolders": {
      "get": {
        "tags": [
          "Folders"
        ],
        "operationId": "GetParentFolders",
        "parameters": [
          {
            "name": "folderId",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FolderDtoApiListResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/Sites/GetByUrl": {
      "get": {
        "tags": [
          "Sites"
        ],
        "operationId": "GetByUrl",
        "parameters": [
          {
            "name": "url",
            "in": "query",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SiteDtoApiResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApiError": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "code": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "FileDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "siteId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "normalizedName": {
            "type": "string",
            "nullable": true
          },
          "folderId": {
            "type": "string",
            "format": "uuid"
          },
          "extension": {
            "type": "string",
            "nullable": true
          },
          "contentType": {
            "type": "string",
            "nullable": true
          },
          "size": {
            "type": "integer",
            "format": "int64"
          }
        }
      },
...
      "FolderDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "siteId": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "normalizedName": {
            "type": "string",
            "nullable": true
          },
          "files": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FileDto"
            }
          },
          "folders": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FolderDto"
            }
          },
          "parentFolder": {
            "$ref": "#/components/schemas/FolderDto"
          }
        }
      },
      "FolderDtoApiListResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageCode": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "uniqueId": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/FolderDto"
            }
          },
          "pagination": {
            "$ref": "#/components/schemas/PaginationInfo"
          }
        }
      },
      "FolderDtoApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageCode": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "uniqueId": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "$ref": "#/components/schemas/FolderDto"
          }
        }
      },
...
      "PaginationInfo": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "page": {
            "type": "integer",
            "format": "int32"
          },
          "pageSize": {
            "type": "integer",
            "format": "int32"
          },
          "totalPages": {
            "type": "integer",
            "format": "int32"
          },
          "totalCount": {
            "type": "integer",
            "format": "int32"
          },
          "hasNextPage": {
            "type": "boolean"
          },
          "hasPreviousPage": {
            "type": "boolean"
          }
        }
      },
...
      "SiteDto": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "urls": {
            "type": "array",
            "nullable": true,
            "items": {
              "type": "string"
            }
          },
          "layoutId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "detailLayoutId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "editLayoutId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "metaTitle": {
            "type": "string",
            "nullable": true
          },
          "metaDescription": {
            "type": "string",
            "nullable": true
          },
          "robotsIndex": {
            "type": "boolean",
            "nullable": true
          },
          "robotsFollow": {
            "type": "boolean",
            "nullable": true
          },
          "robotsTxt": {
            "type": "string",
            "nullable": true
          },
          "googleTagsId": {
            "type": "string",
            "nullable": true
          },
          "ogType": {
            "type": "string",
            "nullable": true
          },
          "head": {
            "type": "string",
            "nullable": true
          }
        }
      },
...
      "SiteDtoApiResponse": {
        "type": "object",
        "additionalProperties": false,
        "properties": {
          "success": {
            "type": "boolean"
          },
          "message": {
            "type": "string",
            "nullable": true
          },
          "messageCode": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "array",
            "nullable": true,
            "items": {
              "$ref": "#/components/schemas/ApiError"
            }
          },
          "timestamp": {
            "type": "string",
            "format": "date-time"
          },
          "traceId": {
            "type": "string",
            "nullable": true
          },
          "statusCode": {
            "type": "integer",
            "format": "int32"
          },
          "sessionId": {
            "type": "string",
            "nullable": true
          },
          "uniqueId": {
            "type": "string",
            "nullable": true
          },
          "duration": {
            "type": "number",
            "format": "double"
          },
          "data": {
            "$ref": "#/components/schemas/SiteDto"
          }
        }
      },
...
    },
    "securitySchemes": {
      "Bearer": {
        "type": "http",
        "description": "JWT Authorization header using the Bearer scheme. Example: \"Bearer {token}\"",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  }
}